% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_functions.R
\name{linear_demand}
\alias{linear_demand}
\title{Linear Demand}
\usage{
linear_demand(a, b, p, beta = NULL, X = NULL)
}
\arguments{
\item{a}{(positive \code{number})\cr intercept of the demand function.}

\item{b}{(negative \code{number})\cr coefficient for price.}

\item{p}{(non-negative \code{number})\cr price.}

\item{beta}{(named \code{numeric})\cr named values of the covariate coefficients.}

\item{X}{(named \code{numeric})\cr named values of the ccovariates.}
}
\value{
This returns the (continuous) linear demand (i.e., a quantity).
}
\description{
Linear Demand
}
\examples{
linear_demand(100, -0.5, 30, c("beta1" = -2), c("x1" = 1))
}
