% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{priceEngine-package}
\alias{priceEngine}
\alias{priceEngine-package}
\title{\code{priceEngine} Package}
\description{
\code{priceEngine} is a package used for dynamic real-world-demand price optimization.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
